ml/requirements.txt
fastapi
uvicorn
paddleocr
pytesseract
google-generativeai
python-multipart
transformers==4.40.0
torch
Pillow
accelerate
bitsandbytes